requirements.txt
django>=5.1
django-htmx>=1.19
django-compressor>=4.5
django-libsass>=0.9
pymongo[srv]>=4.8
gunicorn>=22.0
uvicorn[standard]>=0.29
whitenoise>=6.7
markdown>=3.6
autogen-agentchat>=0.4
requests>=2.32
autogen-core>=0.4
autogen-ext[openai,azure,anthropic]>=0.4
python-json-logger>=2.0
opentelemetry-api>=1.27
opentelemetry-sdk>=1.27
opentelemetry-exporter-otlp-proto-http>=1.27
